pyproject.toml
[tool.poetry]
name = "photo-ticket-api"
version = "0.1.0"
description = ""
authors = ["Juan Ortega"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"

[tool.poetry.dev-dependencies]
black = "^24.4.2"
pre-commit = "^2.19.0"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.pre-commit.repos]]
repo = "https://github.com/psf/black"
rev = "24.4.2"
hooks = [
    {id = "black", args = ["--line-length=88"]}
]
